reorganize test directory
