+++ pyloo/tests/conftest.py
@@ -4,14 +4,97 @@ import os
 
 import arviz as az
 import numpy as np
-import pandas as pd
-import pymc as pm
-import pytensor.tensor as pt
 import pytest
-import xarray as xr
-from arviz import InferenceData
-from pymc_marketing.mmm import MMM, GeometricAdstock, LogisticSaturation
-from pymc_marketing.prior import Prior
+
+from .helpers import data_random, models, multidim_models
+from .models import (
+    approximate_posterior_model,
+    bernoulli_model,
+    hierarchical_model,
+    hierarchical_model_no_coords,
+    high_dimensional_regression_model,
+    large_regression_model,
+    mixture_model,
+    mmm_model,
+    multi_observed_model,
+    mvn_spatial_model,
+    mvt_spatial_model,
+    poisson_model,
+    problematic_k_model,
+    roaches_model,
+    shared_variable_model,
+    simple_model,
+    student_t_model,
+    wells_model,
+)
+from .test_data import (
+    both_cov_prec_data,
+    loo_predictive_metric_binary_data,
+    loo_predictive_metric_data,
+    missing_cov_data,
+    mvn_custom_names_data,
+    mvn_inference_data,
+    mvn_precision_data,
+    mvn_validation_data,
+    mvt_custom_names_data,
+    mvt_inference_data,
+    mvt_negative_df_data,
+    mvt_precision_data,
+    mvt_validation_data,
+    prepare_inference_data_for_crps,
+    singular_matrix_data,
+)
+
+__all__ = [
+    "data_random",
+    "models",
+    "multidim_models",
+    "mmm_model",
+    "hierarchical_model",
+    "hierarchical_model_no_coords",
+    "simple_model",
+    "poisson_model",
+    "multi_observed_model",
+    "shared_variable_model",
+    "bernoulli_model",
+    "large_regression_model",
+    "student_t_model",
+    "mixture_model",
+    "problematic_k_model",
+    "roaches_model",
+    "approximate_posterior_model",
+    "wells_model",
+    "high_dimensional_regression_model",
+    "mvn_spatial_model",
+    "mvt_spatial_model",
+    "loo_predictive_metric_data",
+    "loo_predictive_metric_binary_data",
+    "prepare_inference_data_for_crps",
+    "mvn_inference_data",
+    "mvt_inference_data",
+    "mvn_precision_data",
+    "mvt_precision_data",
+    "mvn_custom_names_data",
+    "mvt_custom_names_data",
+    "mvt_negative_df_data",
+    "singular_matrix_data",
+    "both_cov_prec_data",
+    "missing_cov_data",
+    "mvn_validation_data",
+    "mvt_validation_data",
+    "rng",
+    "centered_eight",
+    "non_centered_eight",
+    "pytest_addoption",
+    "save_dir",
+    "numpy_arrays",
+    "log_likelihood_data",
+    "multidim_data",
+    "extreme_data",
+    "eight_schools_params",
+    "draws",
+    "chains",
+]
 
 
 @pytest.fixture(scope="session")
@@ -107,1436 +190,3 @@ def draws():
 def chains():
     """Share default chain count."""
     return 2
-
-
-@pytest.fixture(scope="session")
-def mmm_model():
-    """Create a media mix model for testing."""
-    rng = np.random.default_rng(42)
-
-    n_weeks = 52 * 3
-
-    x1 = rng.uniform(0, 1, size=n_weeks)
-    x2 = rng.uniform(0, 1, size=n_weeks)
-    x1[x1 > 0.9] *= 2
-    x2[x2 < 0.2] = 0
-
-    t = np.arange(n_weeks)
-    dayofyear = np.linspace(0, 365, n_weeks)
-    seasonality = 0.5 * (
-        np.sin(2 * np.pi * dayofyear / 365) + np.cos(2 * np.pi * dayofyear / 365)
-    )
-
-    event_1 = np.zeros(n_weeks)
-    event_1[20] = 1
-
-    beta_1, beta_2 = 2.0, 1.5
-    trend_coef = 0.02
-    event_effect = 2.0
-
-    y = (
-        1.0
-        + trend_coef * t
-        + seasonality
-        + event_effect * event_1
-        + beta_1 * x1
-        + beta_2 * x2
-        + rng.normal(0, 0.2, size=n_weeks)
-    )
-
-    data = pd.DataFrame({
-        "date_week": pd.date_range(start="2023-01-01", periods=n_weeks, freq="W-MON"),
-        "x1": x1,
-        "x2": x2,
-        "t": t,
-        "event_1": event_1,
-        "y": y,
-    })
-
-    model_config = {
-        "intercept": Prior("Normal", mu=0.0, sigma=1.0),
-        "saturation_beta": Prior("HalfNormal", sigma=[1.0, 1.0]),
-        "gamma_control": Prior("Normal", mu=0, sigma=0.5),
-        "gamma_fourier": Prior("Laplace", mu=0, b=0.2),
-        "likelihood": Prior("Normal", sigma=Prior("HalfNormal", sigma=1.0)),
-    }
-
-    mmm = MMM(
-        model_config=model_config,
-        date_column="date_week",
-        adstock=GeometricAdstock(l_max=4),
-        saturation=LogisticSaturation(),
-        channel_columns=["x1", "x2"],
-        control_columns=["event_1", "t"],
-        yearly_seasonality=1,
-    )
-
-    X = data.drop("y", axis=1)
-    y = data["y"]
-
-    idata = mmm.fit(
-        X=X,
-        y=y,
-        chains=2,
-        draws=500,
-        tune=500,
-        target_accept=0.9,
-        random_seed=42,
-        idata_kwargs={"log_likelihood": True},
-    )
-
-    return mmm.model, idata
-
-
-@pytest.fixture(scope="session")
-def hierarchical_model():
-    """Create a hierarchical model with multiple observations for testing."""
-    rng = np.random.default_rng(42)
-
-    n_groups = 8
-    n_points = 20
-
-    alpha = 0.8
-    beta = 1.2
-    group_effects = rng.normal(0, 0.5, size=n_groups)
-
-    X = rng.normal(0, 1, size=(n_groups, n_points))
-
-    Y = (
-        alpha
-        + group_effects[:, None]
-        + beta * X
-        + rng.normal(0, 0.2, size=(n_groups, n_points))
-    )
-
-    coords = {"group": range(n_groups), "obs_id": range(n_points)}
-
-    with pm.Model(coords=coords) as model:
-        alpha = pm.Normal("alpha", mu=0, sigma=2)
-        beta = pm.Normal("beta", mu=0, sigma=2)
-        group_sigma = pm.HalfNormal("group_sigma", sigma=0.5)
-
-        group_effects_raw = pm.Normal("group_effects_raw", mu=0, sigma=1, dims="group")
-        group_effects = pm.Deterministic(
-            "group_effects", group_effects_raw * group_sigma, dims="group"
-        )
-
-        mu = alpha + group_effects[:, None] + beta * X
-
-        sigma_y = pm.HalfNormal("sigma_y", sigma=0.5)
-        pm.Normal("Y", mu=mu, sigma=sigma_y, observed=Y, dims=("group", "obs_id"))
-
-        idata = pm.sample(
-            1000,
-            tune=2000,
-            target_accept=0.95,
-            random_seed=42,
-            idata_kwargs={"log_likelihood": True},
-        )
-
-    return model, idata
-
-
-@pytest.fixture(scope="session")
-def hierarchical_model_no_coords():
-    """Create a hierarchical model without explicit coordinates."""
-    rng = np.random.default_rng(42)
-
-    n_groups = 8
-    n_points = 20
-
-    alpha = 0.8
-    beta = 1.2
-    group_effects = rng.normal(0, 0.5, size=n_groups)
-
-    X = rng.normal(0, 1, size=(n_groups, n_points))
-    Y = (
-        alpha
-        + group_effects[:, None]
-        + beta * X
-        + rng.normal(0, 0.2, size=(n_groups, n_points))
-    )
-
-    with pm.Model() as model:
-        alpha = pm.Normal("alpha", mu=0, sigma=2)
-        beta = pm.Normal("beta", mu=0, sigma=2)
-        group_sigma = pm.HalfNormal("group_sigma", sigma=0.5)
-
-        group_effects_raw = pm.Normal(
-            "group_effects_raw", mu=0, sigma=1, shape=n_groups
-        )
-        group_effects = pm.Deterministic(
-            "group_effects", group_effects_raw * group_sigma
-        )
-
-        mu = alpha + group_effects[:, None] + beta * X
-
-        sigma_y = pm.HalfNormal("sigma_y", sigma=0.5)
-        pm.Normal("Y", mu=mu, sigma=sigma_y, observed=Y)
-
-        idata = pm.sample(
-            1000,
-            tune=2000,
-            target_accept=0.95,
-            random_seed=42,
-            idata_kwargs={"log_likelihood": True},
-        )
-
-    return model, idata
-
-
-@pytest.fixture(scope="session")
-def simple_model():
-    """Create a simple linear regression model for testing."""
-    rng = np.random.default_rng(42)
-    X = rng.normal(0, 1, size=100)
-    true_alpha = 1.0
-    true_beta = 2.0
-    true_sigma = 1.0
-    y = true_alpha + true_beta * X + rng.normal(0, true_sigma, size=100)
-
-    with pm.Model(coords={"obs_id": range(len(X))}) as model:
-        alpha = pm.Normal("alpha", mu=0, sigma=10)
-        beta = pm.Normal("beta", mu=0, sigma=10)
-        sigma = pm.HalfNormal("sigma", sigma=10)
-
-        mu = alpha + beta * X
-
-        pm.Normal("y", mu=mu, sigma=sigma, observed=y, dims="obs_id")
-
-        idata = pm.sample(
-            1000, tune=1000, random_seed=42, idata_kwargs={"log_likelihood": True}
-        )
-
-    return model, idata
-
-
-@pytest.fixture(scope="session")
-def poisson_model():
-    """Create a Poisson regression model for testing different likelihood types."""
-    rng = np.random.default_rng(42)
-    X = rng.normal(0, 1, size=100)
-    true_alpha = 0.5
-    true_beta = 0.3
-
-    lambda_rate = np.exp(true_alpha + true_beta * X)
-    y = rng.poisson(lambda_rate)
-
-    with pm.Model(coords={"obs_id": range(len(X))}) as model:
-        alpha = pm.Normal("alpha", mu=0, sigma=1)
-        beta = pm.Normal("beta", mu=0, sigma=1)
-
-        lambda_rate = pm.math.exp(alpha + beta * X)
-        pm.Poisson("y", mu=lambda_rate, observed=y, dims="obs_id")
-
-        idata = pm.sample(
-            1000, tune=1000, random_seed=42, idata_kwargs={"log_likelihood": True}
-        )
-
-    return model, idata
-
-
-@pytest.fixture(scope="session")
-def multi_observed_model():
-    """Create a model with multiple observed variables."""
-    rng = np.random.default_rng(42)
-    n_samples = 100
-
-    X = rng.normal(0, 1, size=n_samples)
-    true_alpha = 1.0
-    true_beta = 2.0
-    true_sigma1 = 0.5
-    true_sigma2 = 0.8
-
-    y1 = true_alpha + true_beta * X + rng.normal(0, true_sigma1, size=n_samples)
-    y2 = true_alpha - true_beta * X + rng.normal(0, true_sigma2, size=n_samples)
-
-    with pm.Model(coords={"obs_id": range(n_samples)}) as model:
-        alpha = pm.Normal("alpha", mu=0, sigma=10)
-        beta = pm.Normal("beta", mu=0, sigma=10)
-        sigma1 = pm.HalfNormal("sigma1", sigma=10)
-        sigma2 = pm.HalfNormal("sigma2", sigma=10)
-
-        mu1 = alpha + beta * X
-        mu2 = alpha - beta * X
-
-        pm.Normal("y1", mu=mu1, sigma=sigma1, observed=y1, dims="obs_id")
-        pm.Normal("y2", mu=mu2, sigma=sigma2, observed=y2, dims="obs_id")
-
-        idata = pm.sample(
-            1000, tune=1000, random_seed=42, idata_kwargs={"log_likelihood": True}
-        )
-
-    return model, idata
-
-
-@pytest.fixture(scope="session")
-def shared_variable_model():
-    """Create a model with shared variables across observations."""
-    rng = np.random.default_rng(42)
-    n_groups = 3
-    n_per_group = 50
-
-    shared_effect = 0.5
-    group_effects = rng.normal(0, 0.3, size=n_groups)
-
-    X = []
-    y = []
-    groups = []
-
-    for i in range(n_groups):
-        X_group = rng.normal(0, 1, size=n_per_group)
-        y_group = (
-            shared_effect
-            + group_effects[i]
-            + 0.2 * X_group
-            + rng.normal(0, 0.1, size=n_per_group)
-        )
-
-        X.extend(X_group)
-        y.extend(y_group)
-        groups.extend([i] * n_per_group)
-
-    X = np.array(X)
-    y = np.array(y)
-    groups = np.array(groups)
-
-    coords = {"obs_id": range(len(X)), "group": range(n_groups)}
-
-    with pm.Model(coords=coords) as model:
-        shared_effect = pm.Normal("shared_effect", mu=0, sigma=1)
-        group_effects = pm.Normal("group_effects", mu=0, sigma=0.5, dims="group")
-        beta = pm.Normal("beta", mu=0, sigma=1)
-        sigma = pm.HalfNormal("sigma", sigma=0.5)
-
-        mu = shared_effect + group_effects[groups] + beta * X
-        pm.Normal("y", mu=mu, sigma=sigma, observed=y, dims="obs_id")
-
-        idata = pm.sample(
-            1000, tune=1000, random_seed=42, idata_kwargs={"log_likelihood": True}
-        )
-
-    return model, idata
-
-
-@pytest.fixture(scope="session")
-def bernoulli_model():
-    """Create a Bernoulli model for testing different likelihood types."""
-    rng = np.random.default_rng(42)
-    X = rng.normal(0, 1, size=100)
-    true_alpha = 0.5
-    true_beta = 0.3
-
-    p = 1 / (1 + np.exp(-(true_alpha + true_beta * X)))
-    y = rng.binomial(1, p)
-
-    with pm.Model(coords={"obs_id": range(len(X))}) as model:
-        alpha = pm.Normal("alpha", mu=0, sigma=1)
-        beta = pm.Normal("beta", mu=0, sigma=1)
-
-        p = pm.math.invlogit(alpha + beta * X)
-        pm.Bernoulli("y", p=p, observed=y, dims="obs_id")
-
-        idata = pm.sample(
-            1000, tune=1000, random_seed=42, idata_kwargs={"log_likelihood": True}
-        )
-
-    return model, idata
-
-
-@pytest.fixture(scope="session")
-def large_regression_model():
-    """Create a large linear regression model with many observations for K-fold testing."""
-    rng = np.random.default_rng(42)
-    n_obs = 1000
-    n_predictors = 3
-
-    X = rng.normal(0, 1, size=(n_obs, n_predictors))
-
-    true_alpha = 1.5
-    true_betas = np.array([0.8, -0.5, 1.2])
-    true_sigma = 0.7
-
-    y = true_alpha + X @ true_betas + rng.normal(0, true_sigma, size=n_obs)
-
-    with pm.Model(
-        coords={"obs_id": range(n_obs), "predictor": range(n_predictors)}
-    ) as model:
-        alpha = pm.Normal("alpha", mu=0, sigma=5)
-        betas = pm.Normal("betas", mu=0, sigma=2, dims="predictor")
-        sigma = pm.HalfNormal("sigma", sigma=3)
-
-        mu = alpha + pm.math.dot(X, betas)
-
-        pm.Normal("y", mu=mu, sigma=sigma, observed=y, dims="obs_id")
-
-        idata = pm.sample(
-            draws=1000, tune=500, random_seed=42, idata_kwargs={"log_likelihood": True}
-        )
-
-    return model, idata
-
-
-@pytest.fixture(scope="session")
-def student_t_model():
-    """Create a Student's T model for testing different likelihood types."""
-    rng = np.random.default_rng(42)
-    X = rng.normal(0, 1, size=100)
-    true_alpha = 1.0
-    true_beta = 2.0
-    true_sigma = 1.0
-    true_nu = 4.0
-
-    y = true_alpha + true_beta * X + rng.standard_t(true_nu) * true_sigma
-
-    with pm.Model(coords={"obs_id": range(len(X))}) as model:
-        alpha = pm.Normal("alpha", mu=0, sigma=10)
-        beta = pm.Normal("beta", mu=0, sigma=10)
-        sigma = pm.HalfNormal("sigma", sigma=10)
-        nu = pm.Exponential("nu", lam=1 / 10)
-
-        mu = alpha + beta * X
-        pm.StudentT("y", nu=nu, mu=mu, sigma=sigma, observed=y, dims="obs_id")
-
-        idata = pm.sample(
-            1000, tune=1000, random_seed=42, idata_kwargs={"log_likelihood": True}
-        )
-
-    return model, idata
-
-
-@pytest.fixture(scope="session")
-def mixture_model():
-    """Create a mixture model for testing mixture likelihoods."""
-    rng = np.random.default_rng(42)
-    n_samples = 200
-
-    true_w = 0.7
-    true_mu1 = -2.0
-    true_mu2 = 3.0
-    true_sigma1 = 0.5
-    true_sigma2 = 1.0
-
-    component = rng.binomial(1, true_w, size=n_samples)
-    y = np.zeros(n_samples)
-    y[component == 1] = rng.normal(true_mu1, true_sigma1, size=np.sum(component == 1))
-    y[component == 0] = rng.normal(true_mu2, true_sigma2, size=np.sum(component == 0))
-
-    with pm.Model(coords={"obs_id": range(n_samples)}) as model:
-        w = pm.Beta("w", alpha=2, beta=2)
-
-        mu1 = pm.Normal("mu1", mu=0, sigma=5)
-        mu2 = pm.Normal("mu2", mu=0, sigma=5)
-
-        sigma1 = pm.HalfNormal("sigma1", sigma=2)
-        sigma2 = pm.HalfNormal("sigma2", sigma=2)
-
-        comp1 = pm.Normal.dist(mu=mu1, sigma=sigma1)
-        comp2 = pm.Normal.dist(mu=mu2, sigma=sigma2)
-
-        pm.Mixture(
-            "y", w=[w, 1 - w], comp_dists=[comp1, comp2], observed=y, dims="obs_id"
-        )
-
-        idata = pm.sample(
-            1000,
-            tune=1000,
-            target_accept=0.9,
-            random_seed=42,
-            idata_kwargs={"log_likelihood": True},
-        )
-
-    return model, idata
-
-
-@pytest.fixture(scope="session")
-def problematic_k_model():
-    """Create a model that generates problematic pareto k values for LOO-CV testing."""
-    rng = np.random.default_rng(42)
-
-    n_obs = 2000
-    true_alpha = 1.0
-    true_beta = 2.0
-    true_sigma = 0.1
-    df = 2
-
-    X = rng.normal(0, 1, size=n_obs)
-
-    outlier_indices = np.arange(0, n_obs, 10)
-    X[outlier_indices] = rng.normal(5, 2, size=len(outlier_indices))
-
-    noise = np.zeros(n_obs)
-    regular_indices = np.ones(n_obs, dtype=bool)
-    regular_indices[outlier_indices] = False
-
-    noise[regular_indices] = rng.normal(0, true_sigma, size=regular_indices.sum())
-    noise[outlier_indices] = (
-        rng.standard_t(df, size=len(outlier_indices)) * true_sigma * 10
-    )
-
-    extreme_indices = np.arange(5, n_obs, 50)
-    noise[extreme_indices] = (
-        rng.standard_t(df, size=len(extreme_indices)) * true_sigma * 20
-    )
-
-    y = true_alpha + true_beta * X + noise
-
-    with pm.Model(coords={"obs_id": range(n_obs)}) as model:
-        alpha = pm.Normal("alpha", mu=0, sigma=10)
-        beta = pm.Normal("beta", mu=0, sigma=10)
-        sigma = pm.HalfNormal("sigma", sigma=10)
-
-        mu = alpha + beta * X
-        pm.Normal("y", mu=mu, sigma=sigma, observed=y, dims="obs_id")
-
-        idata = pm.sample(
-            1000,
-            tune=2000,
-            target_accept=0.95,
-            random_seed=42,
-            idata_kwargs={"log_likelihood": True},
-        )
-
-    return model, idata
-
-
-@pytest.fixture(scope="session")
-def roaches_model():
-    """Create a model for the roaches dataset."""
-    file_path = os.path.join(
-        os.path.dirname(os.path.dirname(os.path.dirname(__file__))),
-        "data",
-        "roaches.csv",
-    )
-    data = pd.read_csv(file_path)
-    data["roach1"] = np.sqrt(data["roach1"])
-
-    X = data[["roach1", "treatment", "senior"]].values
-    y = data["y"].values
-    offset = np.log(data["exposure2"]).values
-    K = X.shape[1]
-
-    beta_prior_scale = 2.5
-    alpha_prior_scale = 5.0
-
-    with pm.Model() as model:
-        beta = pm.Normal("beta", mu=0, sigma=beta_prior_scale, shape=K)
-        intercept = pm.Normal("intercept", mu=0, sigma=alpha_prior_scale)
-        eta = pm.math.dot(X, beta) + intercept + offset
-        pm.Poisson("y", mu=pm.math.exp(eta), observed=y)
-
-        idata = pm.sample(
-            chains=4,
-            draws=1000,
-            tune=1000,
-            target_accept=0.9,
-            random_seed=42,
-            idata_kwargs={"log_likelihood": True},
-        )
-
-    return model, idata
-
-
-@pytest.fixture(scope="session")
-def approximate_posterior_model():
-    """Create a model with an approximate posterior (ADVI) for testing."""
-    rng = np.random.default_rng(42)
-    n_samples = 5000
-
-    X = rng.normal(0, 1, size=n_samples)
-    true_alpha = 0.5
-    true_beta = 1.5
-    true_sigma = 0.8
-    y = true_alpha + true_beta * X + rng.normal(0, true_sigma, size=n_samples)
-
-    with pm.Model(coords={"obs_id": range(n_samples)}) as model:
-        alpha = pm.Normal("alpha", mu=0, sigma=2)
-        beta = pm.Normal("beta", mu=0, sigma=2)
-        sigma = pm.HalfNormal("sigma", sigma=2)
-
-        mu = alpha + beta * X
-        pm.Normal("y", mu=mu, sigma=sigma, observed=y, dims="obs_id")
-
-        approx = pm.fit(
-            method="advi",
-            n=1000,
-            random_seed=42,
-        )
-
-        idata = approx.sample(1000, return_inferencedata=True)
-        pm.compute_log_likelihood(idata, model=model, extend_inferencedata=True)
-
-    return model, idata, approx
-
-
-@pytest.fixture(scope="session")
-def wells_model():
-    """Create a logistic regression model for the arsenic wells dataset."""
-    file_path = os.path.join(
-        os.path.dirname(os.path.dirname(os.path.dirname(__file__))), "data", "wells.csv"
-    )
-    data = pd.read_csv(file_path)
-
-    y = data["switch"].values
-
-    data["dist100"] = data["dist"] / 100
-
-    X = np.column_stack([np.ones(len(data)), data[["dist100", "arsenic"]].values])
-
-    P = X.shape[1]
-    N = len(y)
-
-    with pm.Model(coords={"obs_id": range(N), "predictor": range(P)}) as model:
-        beta = pm.Normal("beta", mu=0, sigma=1, dims="predictor")
-
-        eta = pm.math.dot(X, beta)
-
-        pm.Bernoulli("y", logit_p=eta, observed=y, dims="obs_id")
-
-        idata = pm.sample(
-            chains=4,
-            draws=1000,
-            tune=1000,
-            target_accept=0.9,
-            random_seed=42,
-            idata_kwargs={"log_likelihood": True},
-        )
-
-    return model, idata
-
-
-@pytest.fixture(scope="session")
-def loo_predictive_metric_data():
-    """Create test data for loo_predictive_metric tests."""
-    n_chains = 4
-    n_draws = 250
-    n_obs = 10
-
-    rng = np.random.default_rng(42)
-    y = rng.normal(0, 1, n_obs)
-
-    mu_samples = rng.normal(y, 0.1, (n_chains, n_draws, n_obs))
-    log_lik = -0.5 * np.log(2 * np.pi) - 0.5 * ((y - mu_samples) / 1.0) ** 2
-
-    pp_data = xr.Dataset(
-        {"predictions": (["chain", "draw", "observation"], mu_samples)},
-        coords={
-            "chain": np.arange(n_chains),
-            "draw": np.arange(n_draws),
-            "observation": np.arange(n_obs),
-        },
-    )
-
-    ll_data = xr.Dataset(
-        {"obs": (["chain", "draw", "observation"], log_lik)},
-        coords={
-            "chain": np.arange(n_chains),
-            "draw": np.arange(n_draws),
-            "observation": np.arange(n_obs),
-        },
-    )
-
-    obs_data = xr.Dataset(
-        {"obs": (["observation"], y)}, coords={"observation": np.arange(n_obs)}
-    )
-
-    idata = InferenceData(
-        posterior_predictive=pp_data, log_likelihood=ll_data, observed_data=obs_data
-    )
-
-    pp_samples = idata.posterior_predictive.predictions.stack(
-        __sample__=("chain", "draw")
-    )
-    log_lik_samples = idata.log_likelihood.obs.stack(__sample__=("chain", "draw"))
-
-    return {
-        "idata": idata,
-        "y": y,
-        "pp_samples": pp_samples,
-        "log_lik_samples": log_lik_samples,
-    }
-
-
-@pytest.fixture(scope="session")
-def loo_predictive_metric_binary_data():
-    """Create binary test data for loo_predictive_metric tests."""
-    n_chains = 4
-    n_draws = 250
-    n_obs = 100
-
-    rng = np.random.default_rng(42)
-    true_probs = rng.uniform(0.1, 0.9, n_obs)
-    y_binary = rng.binomial(1, true_probs, n_obs)
-
-    y_binary_reshaped = y_binary.reshape(1, 1, -1)
-
-    prob_samples = np.zeros((n_chains, n_draws, n_obs))
-    for i in range(n_chains):
-        for j in range(n_draws):
-            prob_samples[i, j] = rng.beta(y_binary + 1, 2 - y_binary)
-
-    log_lik_binary = y_binary_reshaped * np.log(prob_samples) + (
-        1 - y_binary_reshaped
-    ) * np.log(1 - prob_samples)
-
-    pp_data = xr.Dataset(
-        {"predictions": (["chain", "draw", "observation"], prob_samples)},
-        coords={
-            "chain": np.arange(n_chains),
-            "draw": np.arange(n_draws),
-            "observation": np.arange(n_obs),
-        },
-    )
-
-    ll_data = xr.Dataset(
-        {"obs": (["chain", "draw", "observation"], log_lik_binary)},
-        coords={
-            "chain": np.arange(n_chains),
-            "draw": np.arange(n_draws),
-            "observation": np.arange(n_obs),
-        },
-    )
-
-    obs_data = xr.Dataset(
-        {"obs": (["observation"], y_binary)}, coords={"observation": np.arange(n_obs)}
-    )
-
-    idata = InferenceData(
-        posterior_predictive=pp_data, log_likelihood=ll_data, observed_data=obs_data
-    )
-
-    pp_samples = idata.posterior_predictive.predictions.stack(
-        __sample__=("chain", "draw")
-    )
-    log_lik_samples = idata.log_likelihood.obs.stack(__sample__=("chain", "draw"))
-
-    return {
-        "idata": idata,
-        "y_binary": y_binary,
-        "pp_samples": pp_samples,
-        "log_lik_samples": log_lik_samples,
-    }
-
-
-@pytest.fixture(scope="session")
-def prepare_inference_data_for_crps(centered_eight):
-    """Prepare inference data for CRPS calculations using real model data."""
-    idata = centered_eight
-    pp_orig = idata.posterior_predictive.obs.values
-
-    rng = np.random.default_rng(42)
-    pp_new = pp_orig + rng.normal(0, 0.1, size=pp_orig.shape)
-
-    pp_combined = xr.Dataset(
-        {
-            "obs": (["chain", "draw", "school"], pp_orig),
-            "obs2": (["chain", "draw", "school"], pp_new),
-        },
-        coords=idata.posterior_predictive.coords,
-        attrs=idata.posterior_predictive.attrs,
-    )
-
-    new_idata = InferenceData(
-        posterior=idata.posterior,
-        posterior_predictive=pp_combined,
-        log_likelihood=idata.log_likelihood,
-        observed_data=idata.observed_data,
-    )
-
-    assert hasattr(new_idata, "posterior_predictive")
-    assert "obs" in new_idata.posterior_predictive.data_vars
-    assert "obs2" in new_idata.posterior_predictive.data_vars
-
-    return new_idata
-
-
-@pytest.fixture(scope="session")
-def high_dimensional_regression_model():
-    """Create a high-dimensional linear regression model with many predictors."""
-    rng = np.random.default_rng(42)
-    n_obs = 50
-    n_predictors = 60
-
-    X = rng.normal(0, 1, size=(n_obs, n_predictors))
-
-    true_alpha = 1.5
-    true_betas = rng.normal(0, 1, size=n_predictors)
-    true_sigma = 0.7
-
-    y = true_alpha + X @ true_betas + rng.normal(0, true_sigma, size=n_obs)
-
-    with pm.Model(
-        coords={"obs_id": range(n_obs), "predictor": range(n_predictors)}
-    ) as model:
-        alpha = pm.Normal("alpha", mu=0, sigma=5)
-        betas = pm.Normal("betas", mu=0, sigma=1, dims="predictor")
-        sigma = pm.HalfNormal("sigma", sigma=3)
-
-        mu = alpha + pm.math.dot(X, betas)
-
-        pm.Normal("y", mu=mu, sigma=sigma, observed=y, dims="obs_id")
-
-        idata = pm.sample(
-            draws=1000, tune=500, random_seed=42, idata_kwargs={"log_likelihood": True}
-        )
-
-    return model, idata
-
-
-@pytest.fixture(scope="session")
-def mvn_inference_data():
-    """Create InferenceData with multivariate normal model data."""
-    rng = np.random.default_rng(42)
-    n_obs = 10
-    n_samples = 100
-    n_chains = 2
-
-    mu_samples = rng.normal(0, 1, size=(n_chains, n_samples, n_obs))
-    cov_samples = np.empty((n_chains, n_samples, n_obs, n_obs))
-
-    for c in range(n_chains):
-        for s in range(n_samples):
-            A = rng.normal(0, 0.1, size=(n_obs, n_obs))
-            cov_samples[c, s, :, :] = A @ A.T + np.eye(n_obs) * 0.5
-
-    y_obs_vals = rng.normal(0, 1, size=n_obs)
-
-    idata = InferenceData(
-        posterior=xr.Dataset(
-            {
-                "mu": (("chain", "draw", "obs_dim"), mu_samples),
-                "cov": (("chain", "draw", "obs_dim", "obs_dim_bis"), cov_samples),
-            },
-            coords={
-                "chain": np.arange(n_chains),
-                "draw": np.arange(n_samples),
-                "obs_dim": np.arange(n_obs),
-                "obs_dim_bis": np.arange(n_obs),
-            },
-        ),
-        observed_data=xr.Dataset(
-            {"y": (("obs_dim",), y_obs_vals)},
-            coords={"obs_dim": np.arange(n_obs)},
-        ),
-    )
-
-    return idata
-
-
-@pytest.fixture(scope="session")
-def mvt_inference_data():
-    """Create InferenceData with multivariate Student-t model data."""
-    rng = np.random.default_rng(42)
-    n_obs = 10
-    n_samples = 100
-    n_chains = 2
-
-    mu_samples = rng.normal(0, 1, size=(n_chains, n_samples, n_obs))
-    cov_samples = np.empty((n_chains, n_samples, n_obs, n_obs))
-    df_samples = np.abs(rng.gamma(5, 1, size=(n_chains, n_samples))) + 2.0
-
-    for c in range(n_chains):
-        for s in range(n_samples):
-            A = rng.normal(0, 0.1, size=(n_obs, n_obs))
-            cov_samples[c, s, :, :] = A @ A.T + np.eye(n_obs) * 0.5
-
-    y_obs_vals = rng.normal(0, 1, size=n_obs)
-
-    idata = InferenceData(
-        posterior=xr.Dataset(
-            {
-                "mu": (("chain", "draw", "obs_dim"), mu_samples),
-                "cov": (("chain", "draw", "obs_dim", "obs_dim_bis"), cov_samples),
-                "df": (("chain", "draw"), df_samples),
-            },
-            coords={
-                "chain": np.arange(n_chains),
-                "draw": np.arange(n_samples),
-                "obs_dim": np.arange(n_obs),
-                "obs_dim_bis": np.arange(n_obs),
-            },
-        ),
-        observed_data=xr.Dataset(
-            {"y": (("obs_dim",), y_obs_vals)},
-            coords={"obs_dim": np.arange(n_obs)},
-        ),
-    )
-
-    return idata
-
-
-@pytest.fixture(scope="session")
-def mvn_precision_data():
-    """Create InferenceData with precision matrix instead of covariance."""
-    rng = np.random.default_rng(42)
-    n_obs = 8
-    n_samples = 120
-    n_chains = 2
-
-    mu_samples = rng.normal(0, 1, size=(n_chains, n_samples, n_obs))
-    prec_samples = np.empty((n_chains, n_samples, n_obs, n_obs))
-
-    for c in range(n_chains):
-        for s in range(n_samples):
-            A = rng.normal(0, 0.1, size=(n_obs, n_obs))
-            cov_s = A @ A.T + np.eye(n_obs) * 0.5
-            try:
-                prec_samples[c, s, :, :] = np.linalg.inv(cov_s)
-            except np.linalg.LinAlgError:
-                prec_samples[c, s, :, :] = np.eye(n_obs)
-
-    y_obs_vals = rng.normal(0, 1, size=n_obs)
-
-    idata = InferenceData(
-        posterior=xr.Dataset(
-            {
-                "mu": (("chain", "draw", "obs_dim"), mu_samples),
-                "prec": (("chain", "draw", "obs_dim", "obs_dim_bis"), prec_samples),
-            },
-            coords={
-                "chain": np.arange(n_chains),
-                "draw": np.arange(n_samples),
-                "obs_dim": np.arange(n_obs),
-                "obs_dim_bis": np.arange(n_obs),
-            },
-        ),
-        observed_data=xr.Dataset(
-            {"y": (("obs_dim",), y_obs_vals)},
-            coords={"obs_dim": np.arange(n_obs)},
-        ),
-    )
-
-    return idata
-
-
-@pytest.fixture(scope="session")
-def mvt_precision_data():
-    """Create InferenceData with Student-t model using precision matrix."""
-    rng = np.random.default_rng(42)
-    n_obs = 8
-    n_samples = 120
-    n_chains = 2
-
-    mu_samples = rng.normal(0, 1, size=(n_chains, n_samples, n_obs))
-    prec_samples = np.empty((n_chains, n_samples, n_obs, n_obs))
-    df_samples = np.abs(rng.gamma(5, 1, size=(n_chains, n_samples))) + 2.0
-
-    for c in range(n_chains):
-        for s in range(n_samples):
-            A = rng.normal(0, 0.1, size=(n_obs, n_obs))
-            cov_s = A @ A.T + np.eye(n_obs) * 0.5
-            try:
-                prec_samples[c, s, :, :] = np.linalg.inv(cov_s)
-            except np.linalg.LinAlgError:
-                prec_samples[c, s, :, :] = np.eye(n_obs)
-
-    y_obs_vals = rng.normal(0, 1, size=n_obs)
-
-    idata = InferenceData(
-        posterior=xr.Dataset(
-            {
-                "mu": (("chain", "draw", "obs_dim"), mu_samples),
-                "prec": (("chain", "draw", "obs_dim", "obs_dim_bis"), prec_samples),
-                "df": (("chain", "draw"), df_samples),
-            },
-            coords={
-                "chain": np.arange(n_chains),
-                "draw": np.arange(n_samples),
-                "obs_dim": np.arange(n_obs),
-                "obs_dim_bis": np.arange(n_obs),
-            },
-        ),
-        observed_data=xr.Dataset(
-            {"y": (("obs_dim",), y_obs_vals)},
-            coords={"obs_dim": np.arange(n_obs)},
-        ),
-    )
-
-    return idata
-
-
-@pytest.fixture(scope="session")
-def mvn_custom_names_data():
-    """Create InferenceData with custom variable names."""
-    rng = np.random.default_rng(42)
-    n_obs = 5
-    n_samples = 10
-    n_chains = 2
-
-    mu_samples = rng.normal(0, 1, size=(n_chains, n_samples, n_obs))
-    cov_samples = np.empty((n_chains, n_samples, n_obs, n_obs))
-
-    for c in range(n_chains):
-        for s in range(n_samples):
-            A = rng.normal(0, 0.1, size=(n_obs, n_obs))
-            cov_samples[c, s, :, :] = A @ A.T + np.eye(n_obs) * 0.5
-
-    y_obs_vals = rng.normal(0, 1, size=n_obs)
-
-    idata = InferenceData(
-        posterior=xr.Dataset(
-            {
-                "mean_vector": (("chain", "draw", "obs_dim"), mu_samples),
-                "covariance_matrix": (
-                    ("chain", "draw", "obs_dim", "obs_dim_bis"),
-                    cov_samples,
-                ),
-            },
-            coords={
-                "chain": np.arange(n_chains),
-                "draw": np.arange(n_samples),
-                "obs_dim": np.arange(n_obs),
-                "obs_dim_bis": np.arange(n_obs),
-            },
-        ),
-        observed_data=xr.Dataset(
-            {"observations": (("obs_dim",), y_obs_vals)},
-            coords={"obs_dim": np.arange(n_obs)},
-        ),
-    )
-
-    return idata
-
-
-@pytest.fixture(scope="session")
-def mvt_custom_names_data():
-    """Create InferenceData with Student-t model using custom variable names."""
-    rng = np.random.default_rng(42)
-    n_obs = 5
-    n_samples = 10
-    n_chains = 2
-
-    mu_samples = rng.normal(0, 1, size=(n_chains, n_samples, n_obs))
-    cov_samples = np.empty((n_chains, n_samples, n_obs, n_obs))
-    df_samples = np.abs(rng.gamma(5, 1, size=(n_chains, n_samples))) + 2.0
-
-    for c in range(n_chains):
-        for s in range(n_samples):
-            A = rng.normal(0, 0.1, size=(n_obs, n_obs))
-            cov_samples[c, s, :, :] = A @ A.T + np.eye(n_obs) * 0.5
-
-    y_obs_vals = rng.normal(0, 1, size=n_obs)
-
-    idata = InferenceData(
-        posterior=xr.Dataset(
-            {
-                "location": (("chain", "draw", "obs_dim"), mu_samples),
-                "scale_matrix": (
-                    ("chain", "draw", "obs_dim", "obs_dim_bis"),
-                    cov_samples,
-                ),
-                "nu": (("chain", "draw"), df_samples),
-            },
-            coords={
-                "chain": np.arange(n_chains),
-                "draw": np.arange(n_samples),
-                "obs_dim": np.arange(n_obs),
-                "obs_dim_bis": np.arange(n_obs),
-            },
-        ),
-        observed_data=xr.Dataset(
-            {"observations": (("obs_dim",), y_obs_vals)},
-            coords={"obs_dim": np.arange(n_obs)},
-        ),
-    )
-
-    return idata
-
-
-@pytest.fixture(scope="session")
-def mvt_negative_df_data():
-    """Create InferenceData with Student-t model including negative degrees of freedom."""
-    rng = np.random.default_rng(42)
-    n_obs = 5
-    n_samples = 10
-    n_chains = 2
-
-    mu_samples = rng.normal(0, 1, size=(n_chains, n_samples, n_obs))
-    cov_samples = np.empty((n_chains, n_samples, n_obs, n_obs))
-
-    df_samples = np.abs(rng.gamma(5, 1, size=(n_chains, n_samples))) + 2.0
-    df_samples[0, 0] = -1.0
-
-    for c in range(n_chains):
-        for s in range(n_samples):
-            A = rng.normal(0, 0.1, size=(n_obs, n_obs))
-            cov_samples[c, s, :, :] = A @ A.T + np.eye(n_obs) * 0.5
-
-    y_obs_vals = rng.normal(0, 1, size=n_obs)
-
-    idata = InferenceData(
-        posterior=xr.Dataset(
-            {
-                "mu": (("chain", "draw", "obs_dim"), mu_samples),
-                "cov": (("chain", "draw", "obs_dim", "obs_dim_bis"), cov_samples),
-                "df": (("chain", "draw"), df_samples),
-            },
-            coords={
-                "chain": np.arange(n_chains),
-                "draw": np.arange(n_samples),
-                "obs_dim": np.arange(n_obs),
-                "obs_dim_bis": np.arange(n_obs),
-            },
-        ),
-        observed_data=xr.Dataset(
-            {"y": (("obs_dim",), y_obs_vals)},
-            coords={"obs_dim": np.arange(n_obs)},
-        ),
-    )
-
-    return idata
-
-
-@pytest.fixture(scope="session")
-def singular_matrix_data():
-    """Create InferenceData with singular covariance matrix."""
-    rng = np.random.default_rng(42)
-    n_obs = 5
-    n_samples = 10
-    n_chains = 2
-
-    mu_samples = rng.normal(0, 1, size=(n_chains, n_samples, n_obs))
-    cov_samples = np.empty((n_chains, n_samples, n_obs, n_obs))
-
-    for c in range(n_chains):
-        for s in range(n_samples):
-            if c == 0 and s == 0:
-                cov_samples[c, s, :, :] = np.ones((n_obs, n_obs)) * 0.1
-            else:
-                A = rng.normal(0, 0.1, size=(n_obs, n_obs))
-                cov_samples[c, s, :, :] = A @ A.T + np.eye(n_obs) * 0.5
-
-    y_obs_vals = rng.normal(0, 1, size=n_obs)
-
-    idata = InferenceData(
-        posterior=xr.Dataset(
-            {
-                "mu": (("chain", "draw", "obs_dim"), mu_samples),
-                "cov": (("chain", "draw", "obs_dim", "obs_dim_bis"), cov_samples),
-            },
-            coords={
-                "chain": np.arange(n_chains),
-                "draw": np.arange(n_samples),
-                "obs_dim": np.arange(n_obs),
-                "obs_dim_bis": np.arange(n_obs),
-            },
-        ),
-        observed_data=xr.Dataset(
-            {"y": (("obs_dim",), y_obs_vals)},
-            coords={"obs_dim": np.arange(n_obs)},
-        ),
-    )
-
-    return idata
-
-
-@pytest.fixture(scope="session")
-def both_cov_prec_data():
-    """Create InferenceData with both covariance and precision matrices."""
-    rng = np.random.default_rng(42)
-    n_obs = 5
-    n_samples = 10
-    n_chains = 2
-
-    mu_samples = rng.normal(0, 1, size=(n_chains, n_samples, n_obs))
-    cov_samples = np.empty((n_chains, n_samples, n_obs, n_obs))
-    prec_samples = np.empty((n_chains, n_samples, n_obs, n_obs))
-
-    for c in range(n_chains):
-        for s in range(n_samples):
-            A = rng.normal(0, 0.1, size=(n_obs, n_obs))
-            cov_samples[c, s, :, :] = A @ A.T + np.eye(n_obs) * 0.5
-            prec_samples[c, s, :, :] = np.linalg.inv(cov_samples[c, s, :, :])
-
-    y_obs_vals = rng.normal(0, 1, size=n_obs)
-
-    idata = InferenceData(
-        posterior=xr.Dataset(
-            {
-                "mu": (("chain", "draw", "obs_dim"), mu_samples),
-                "cov": (("chain", "draw", "obs_dim", "obs_dim_bis"), cov_samples),
-                "prec": (("chain", "draw", "obs_dim", "obs_dim_bis"), prec_samples),
-            },
-            coords={
-                "chain": np.arange(n_chains),
-                "draw": np.arange(n_samples),
-                "obs_dim": np.arange(n_obs),
-                "obs_dim_bis": np.arange(n_obs),
-            },
-        ),
-        observed_data=xr.Dataset(
-            {"y": (("obs_dim",), y_obs_vals)},
-            coords={"obs_dim": np.arange(n_obs)},
-        ),
-    )
-
-    return idata
-
-
-@pytest.fixture(scope="session")
-def missing_cov_data():
-    """Create InferenceData with missing covariance matrix."""
-    rng = np.random.default_rng(42)
-    n_obs = 5
-    n_samples = 10
-    n_chains = 2
-
-    mu_samples = rng.normal(0, 1, size=(n_chains, n_samples, n_obs))
-    y_obs_vals = rng.normal(0, 1, size=n_obs)
-
-    idata = InferenceData(
-        posterior=xr.Dataset(
-            {
-                "mu": (("chain", "draw", "obs_dim"), mu_samples),
-            },
-            coords={
-                "chain": np.arange(n_chains),
-                "draw": np.arange(n_samples),
-                "obs_dim": np.arange(n_obs),
-            },
-        ),
-        observed_data=xr.Dataset(
-            {"y": (("obs_dim",), y_obs_vals)},
-            coords={"obs_dim": np.arange(n_obs)},
-        ),
-    )
-
-    return idata
-
-
-@pytest.fixture(scope="session")
-def mvn_validation_data():
-    """Create data for testing model structure validation."""
-    rng = np.random.default_rng(42)
-    n_obs = 5
-    n_samples = 10
-    n_chains = 2
-
-    mu_samples = rng.normal(0, 1, size=(n_chains, n_samples, n_obs))
-    cov_samples = rng.normal(0, 1, size=(n_chains, n_samples, n_obs, n_obs))
-
-    for c in range(n_chains):
-        for s in range(n_samples):
-            cov_samples[c, s] = cov_samples[c, s] @ cov_samples[c, s].T + np.eye(n_obs)
-
-    valid_idata = InferenceData(
-        posterior=xr.Dataset(
-            {
-                "mu": (("chain", "draw", "obs_dim"), mu_samples),
-                "cov": (("chain", "draw", "obs_dim", "obs_dim_bis"), cov_samples),
-            },
-            coords={
-                "chain": np.arange(n_chains),
-                "draw": np.arange(n_samples),
-                "obs_dim": np.arange(n_obs),
-                "obs_dim_bis": np.arange(n_obs),
-            },
-        )
-    )
-
-    no_mu_idata = InferenceData(
-        posterior=xr.Dataset(
-            {
-                "cov": (("chain", "draw", "obs_dim", "obs_dim_bis"), cov_samples),
-            },
-            coords={
-                "chain": np.arange(n_chains),
-                "draw": np.arange(n_samples),
-                "obs_dim": np.arange(n_obs),
-                "obs_dim_bis": np.arange(n_obs),
-            },
-        )
-    )
-
-    no_cov_prec_idata = InferenceData(
-        posterior=xr.Dataset(
-            {
-                "mu": (("chain", "draw", "obs_dim"), mu_samples),
-            },
-            coords={
-                "chain": np.arange(n_chains),
-                "draw": np.arange(n_samples),
-                "obs_dim": np.arange(n_obs),
-            },
-        )
-    )
-
-    no_posterior_idata = InferenceData(
-        observed_data=xr.Dataset(
-            {"y": (("obs_dim",), rng.normal(0, 1, size=n_obs))},
-            coords={"obs_dim": np.arange(n_obs)},
-        )
-    )
-
-    return {
-        "valid": valid_idata,
-        "no_mu": no_mu_idata,
-        "no_cov_prec": no_cov_prec_idata,
-        "no_posterior": no_posterior_idata,
-    }
-
-
-@pytest.fixture(scope="session")
-def mvt_validation_data():
-    """Create data for testing Student-t model structure validation."""
-    rng = np.random.default_rng(42)
-    n_obs = 5
-    n_samples = 10
-    n_chains = 2
-
-    mu_samples = rng.normal(0, 1, size=(n_chains, n_samples, n_obs))
-    cov_samples = rng.normal(0, 1, size=(n_chains, n_samples, n_obs, n_obs))
-    df_samples = np.abs(rng.gamma(5, 1, size=(n_chains, n_samples))) + 2.0
-
-    for c in range(n_chains):
-        for s in range(n_samples):
-            cov_samples[c, s] = cov_samples[c, s] @ cov_samples[c, s].T + np.eye(n_obs)
-
-    valid_idata = InferenceData(
-        posterior=xr.Dataset(
-            {
-                "mu": (("chain", "draw", "obs_dim"), mu_samples),
-                "cov": (("chain", "draw", "obs_dim", "obs_dim_bis"), cov_samples),
-                "df": (("chain", "draw"), df_samples),
-            },
-            coords={
-                "chain": np.arange(n_chains),
-                "draw": np.arange(n_samples),
-                "obs_dim": np.arange(n_obs),
-                "obs_dim_bis": np.arange(n_obs),
-            },
-        )
-    )
-
-    missing_df_idata = InferenceData(
-        posterior=xr.Dataset(
-            {
-                "mu": (("chain", "draw", "obs_dim"), mu_samples),
-                "cov": (("chain", "draw", "obs_dim", "obs_dim_bis"), cov_samples),
-            },
-            coords={
-                "chain": np.arange(n_chains),
-                "draw": np.arange(n_samples),
-                "obs_dim": np.arange(n_obs),
-                "obs_dim_bis": np.arange(n_obs),
-            },
-        )
-    )
-
-    return {"valid": valid_idata, "missing_df": missing_df_idata}
-
-
-@pytest.fixture(scope="session")
-def mvn_spatial_model():
-    """Create a PyMC spatial model with multivariate normal likelihood."""
-    if pm is None:
-        pytest.skip("PyMC not installed")
-
-    np.random.seed(42)
-    n_obs = 25
-
-    coords_x = np.random.uniform(0, 10, size=n_obs)
-    coords_y = np.random.uniform(0, 10, size=n_obs)
-
-    coords = np.vstack([coords_x, coords_y]).T
-    dists = np.zeros((n_obs, n_obs))
-    for i in range(n_obs):
-        for j in range(n_obs):
-            dists[i, j] = np.sqrt(np.sum((coords[i] - coords[j]) ** 2))
-
-    true_sigma = 1.0
-    true_ls = 2.0
-    true_cov = true_sigma**2 * np.exp(-dists / true_ls)
-    true_cov += np.eye(n_obs) * 0.01
-
-    true_mean = 2 + 0.5 * coords_x - 0.3 * coords_y
-    y_obs = np.random.multivariate_normal(true_mean, true_cov)
-
-    with pm.Model() as model:
-        intercept = pm.Normal("intercept", mu=0, sigma=2)
-        beta_x = pm.Normal("beta_x", mu=0, sigma=1)
-        beta_y = pm.Normal("beta_y", mu=0, sigma=1)
-
-        mean_values = intercept + beta_x * coords_x + beta_y * coords_y
-
-        pm.Deterministic("mu", mean_values)
-
-        sigma = pm.HalfNormal("sigma", sigma=2)
-        length_scale = pm.Gamma("length_scale", alpha=2, beta=1)
-
-        cov_values = sigma**2 * pt.exp(-dists / length_scale)
-        cov_values = cov_values + pt.eye(n_obs) * 0.01
-
-        pm.Deterministic("cov", cov_values)
-        pm.MvNormal("y_obs", mu=mean_values, cov=cov_values, observed=y_obs)
-
-        idata = pm.sample(
-            draws=500,
-            tune=1000,
-            chains=2,
-            return_inferencedata=True,
-            target_accept=0.95,
-        )
-
-    return model, idata, coords_x, coords_y, dists
-
-
-@pytest.fixture(scope="session")
-def mvt_spatial_model():
-    """Create a PyMC spatial model with multivariate Student-t likelihood."""
-    if pm is None:
-        pytest.skip("PyMC not installed")
-
-    np.random.seed(42)
-    n_obs = 20
-
-    coords_x = np.random.uniform(0, 10, size=n_obs)
-    coords_y = np.random.uniform(0, 10, size=n_obs)
-
-    coords = np.vstack([coords_x, coords_y]).T
-    dists = np.zeros((n_obs, n_obs))
-    for i in range(n_obs):
-        for j in range(n_obs):
-            dists[i, j] = np.sqrt(np.sum((coords[i] - coords[j]) ** 2))
-
-    true_df = 4.0
-    true_sigma = 1.0
-    true_ls = 2.0
-    true_cov = true_sigma**2 * np.exp(-dists / true_ls)
-    true_cov += np.eye(n_obs) * 0.01
-
-    true_mean = 2 + 0.5 * coords_x - 0.3 * coords_y
-
-    z = np.random.multivariate_normal(np.zeros(n_obs), true_cov)
-    chi2 = np.random.chisquare(true_df) / true_df
-    y_obs = true_mean + z / np.sqrt(chi2)
-
-    with pm.Model() as model:
-        intercept = pm.Normal("intercept", mu=0, sigma=2)
-        beta_x = pm.Normal("beta_x", mu=0, sigma=1)
-        beta_y = pm.Normal("beta_y", mu=0, sigma=1)
-
-        mean_values = intercept + beta_x * coords_x + beta_y * coords_y
-        pm.Deterministic("mu", mean_values)
-
-        sigma = pm.HalfNormal("sigma", sigma=2)
-        length_scale = pm.Gamma("length_scale", alpha=2, beta=1)
-
-        cov_values = sigma**2 * pt.exp(-dists / length_scale)
-        cov_values = cov_values + pt.eye(n_obs) * 0.01
-        pm.Deterministic("cov", cov_values)
-
-        df = pm.Gamma("df", alpha=2, beta=0.5)
-        pm.Deterministic("df_det", df)
-
-        pm.MvStudentT("y_obs", nu=df, mu=mean_values, scale=cov_values, observed=y_obs)
-
-        idata = pm.sample(
-            draws=500,
-            tune=1000,
-            chains=2,
-            return_inferencedata=True,
-            target_accept=0.95,
-        )
-
-    return model, idata, coords_x, coords_y, dists
